pyproject.toml
[project]
name = "ontomem"
version = "0.1.0"
description = "A self-consolidating memory layer for AI agents with schema-first design, intelligent merging, and hybrid search capabilities"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "yifanfeng97", email = "evanfeng97@gmail.com" }]
keywords = ["memory", "ai-agent", "rag", "semantic-search", "llm", "pydantic", "knowledge-graph"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.1",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.6",
    "loguru>=0.7.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]

[project.optional-dependencies]
dev = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.7.1",
    "mkdocstrings[python]>=1.0.0",
    "pytest>=9.0.2",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
